feature(token-linker): renaming contracts
