+++ contracts/nft-linker/NftLinkerBase.sol
@@ -8,7 +8,7 @@ import { AxelarExecutable } from '../executables/AxelarExecutable.sol';
 import { AddressToString, StringToAddress } from '../StringAddressUtils.sol';
 import { Upgradable } from '../upgradables/Upgradable.sol';
 
-abstract contract NftLinker is AxelarExecutable, Upgradable {
+abstract contract NftLinkerBase is AxelarExecutable, Upgradable {
     using StringToAddress for string;
     using AddressToString for address;
