+++ contracts/nft-linker/NftLinkerMintBurn.sol
@@ -2,10 +2,10 @@
 
 pragma solidity ^0.8.0;
 
-import { IERC721 } from '../interfaces/IERC721.sol';
-import { NftLinker } from './NftLinker.sol';
+import { IERC721MintableBurnable } from '../interfaces/IERC721MintableBurnable.sol';
+import { NftLinkerBase } from './NftLinkerBase.sol';
 
-contract NftLinkerLockUnlock is NftLinker {
+contract NftLinkerMintBurn is NftLinkerBase {
     error TransferFailed();
     error TransferFromFailed();
 
@@ -15,22 +15,25 @@ contract NftLinkerLockUnlock is NftLinker {
         address gatewayAddress_,
         address gasServiceAddress_,
         address operatorAddress_
-    ) NftLinker(gatewayAddress_, gasServiceAddress_) {
+    ) NftLinkerBase(gatewayAddress_, gasServiceAddress_) {
         operatorAddress = operatorAddress_;
     }
 
     function _giveNft(address to, uint256 tokenId) internal override {
         (bool success, bytes memory returnData) = operatorAddress.call(
-            abi.encodeWithSelector(IERC721.transferFrom.selector, address(this), to, tokenId)
+            abi.encodeWithSelector(IERC721MintableBurnable.mint.selector, to, tokenId)
         );
         bool transferred = success && (returnData.length == uint256(0) || abi.decode(returnData, (bool)));
 
         if (!transferred || operatorAddress.code.length == 0) revert TransferFailed();
     }
 
-    function _takeNft(address from, uint256 tokenId) internal override {
+    function _takeNft(
+        address, /*from*/
+        uint256 tokenId
+    ) internal override {
         (bool success, bytes memory returnData) = operatorAddress.call(
-            abi.encodeWithSelector(IERC721.transferFrom.selector, from, address(this), tokenId)
+            abi.encodeWithSelector(IERC721MintableBurnable.burn.selector, tokenId)
         );
         bool transferred = success && (returnData.length == uint256(0) || abi.decode(returnData, (bool)));
