+++ contracts/token-linker/TokenLinkerMintBurn.sol
@@ -3,9 +3,9 @@
 pragma solidity ^0.8.0;
 
 import { IERC20MintableBurnable } from '../interfaces/IERC20MintableBurnable.sol';
-import { TokenLinker } from './TokenLinker.sol';
+import { TokenLinkerBase } from './TokenLinkerBase.sol';
 
-contract TokenLinkerMintBurn is TokenLinker {
+contract TokenLinkerMintBurn is TokenLinkerBase {
     error MintFailed();
     error BurnFailed();
 
@@ -15,7 +15,7 @@ contract TokenLinkerMintBurn is TokenLinker {
         address gatewayAddress_,
         address gasServiceAddress_,
         address tokenAddress_
-    ) TokenLinker(gatewayAddress_, gasServiceAddress_) {
+    ) TokenLinkerBase(gatewayAddress_, gasServiceAddress_) {
         tokenAddress = tokenAddress_;
     }
